[WIP] Enable dotty and scala dependencies on the same setup
- dotty dependency is not resolved yet
