+++ project/plugins.sbt
@@ -3,3 +3,5 @@ logLevel := Level.Warn
 
 // sbt-jmh plugin - pulls in JMH dependencies too
 addSbtPlugin("pl.project13.scala" % "sbt-jmh" % "0.2.16")
+
+addSbtPlugin("ch.epfl.lamp" % "sbt-dotty" % "0.1.0-RC2")
